client/requirements.txt
requests>=2.31.0
Pillow>=10.0.0
RPi.GPIO>=0.7.1
